+++ src/dbsolve-export/index.js
@@ -180,15 +180,6 @@ class DBSolveExport extends AbstractExport{
     let discreteEvents = ns
       .selectByClassName('DSwitcher')
       .map((switcher) => {
-        let dynamicRecords = ns
-          .selectRecordsByContext(switcher.id)
-          .filter((record) => record.isDynamic)
-          .map((record) => record.index);
-        if (dynamicRecords.length > 0) {
-          let msg = `DBSolve doesn't support ${switcher.className} for dynamic records: ${dynamicRecords.join(', ')}.`;
-          logger.error(msg, {type: 'ExportError'});
-        }
-        
         // check boolean expression in trigger
         if (!switcher.trigger.isComparison) {
           let msg = `DBSolve supports only simple comparison operators in DSwitcher trigger, got: "${switcher.trigger.toString()}"`;
@@ -203,7 +194,7 @@ class DBSolveExport extends AbstractExport{
               : record.getAssignment(switcher.id);
 
             return {
-              target: record.id + (record.isDynamic ? '_' : ''),
+              targetObj: record,
               expr: expr
             };
           });
@@ -218,15 +209,6 @@ class DBSolveExport extends AbstractExport{
     let continuousEvents = ns
       .selectByClassName('CSwitcher')
       .map((switcher) => {
-        let dynamicRecords = ns
-          .selectRecordsByContext(switcher.id)
-          .filter((record) => record.isDynamic)
-          .map((record) => record.index);
-        if (dynamicRecords.length > 0) {
-          let msg = `DBSolve doesn't support ${switcher.className} for dynamic records: ${dynamicRecords.join(', ')}.`;
-          logger.error(msg, {type: 'ExportError'});
-        }
-
         let assignments = ns
           .selectRecordsByContext(switcher.id)
           .map((record) => {
@@ -235,7 +217,7 @@ class DBSolveExport extends AbstractExport{
               : record.getAssignment(switcher.id);
 
             return {
-              target: record.id + (record.isDynamic ? '_' : ''),
+              targetObj: record,
               expr: expr
             };
           });